Move redirect and notFound to navigation exports
