+++ test/e2e/app-dir/app/app/not-found/client-side/page.js
@@ -1,6 +1,6 @@
 'use client'
 
-import { notFound } from 'next/dist/client/components/not-found'
+import { notFound } from 'next/navigation'
 import React from 'react'
 
 export default function Page() {
