+++ test/e2e/app-dir/app/app/not-found/clientcomponent/client-component.js
@@ -1,5 +1,5 @@
 'use client'
-import { notFound } from 'next/dist/client/components/not-found'
+import { notFound } from 'next/navigation'
 
 export default function ClientComp() {
   notFound()
