+++ test/e2e/app-dir/app/app/not-found/servercomponent/page.js
@@ -1,5 +1,5 @@
 // TODO-APP: enable when flight error serialization is implemented
-import { notFound } from 'next/dist/client/components/not-found'
+import { notFound } from 'next/navigation'
 
 export default function Page() {
   notFound()
