+++ test/e2e/app-dir/app/app/redirect/clientcomponent/client-component.js
@@ -1,5 +1,5 @@
 'use client'
-import { redirect } from 'next/dist/client/components/redirect'
+import { redirect } from 'next/navigation'
 
 export default function ClientComp() {
   redirect('/redirect/result')
