+++ test/e2e/app-dir/app/app/redirect/servercomponent/page.js
@@ -1,4 +1,4 @@
-import { redirect } from 'next/dist/client/components/redirect'
+import { redirect } from 'next/navigation'
 
 export default function Page() {
   redirect('/redirect/result')
